HaplinMethyl: Additional functions for Haplin, for using matrices of environmental data (e.g., DNA methylation data)

The HaplinMethyl package provides three categories of important functions: reading the data, loading the read-in data, and manipulating the environmental data.

Reading the environmental data

The two functions (envDataRead and envDataReadFromObj) read in the data and create an object of env.data class, which is a list of ff-matrices. Simultaneously, two files are written: .RData and .ffData. These are connected and will be used in the data-loading function. The object created here has also summary and print methods defined for easier viewing.

Loading the read-in data

envDataLoad function loads the data, which is super-fast, compared to reading the data in.

Manipulating the data

The user can extract a subset of the data read in by any of the above functions, with envDataSubset. This will create new .RData and .ffData files. Moreover, there are two functions that are created specifically for DNA methylation data: findCpGsnearSNP and findCpGsRange can be used to create CpG lists that are near a given SNP or within a certain coordinate range.